Dev: Proper fix of 17469

+++ application/controllers/admin/participantsaction.php
@@ -1911,7 +1911,8 @@ $url .= "_view"; });
     public function editShareInfo()
     {
         $operation = Yii::app()->request->getPost('oper');
-        $shareIds = (int) Yii::app()->request->getPost('id');
+        // NB: Comma-separated list.
+        $shareIds = Yii::app()->request->getPost('id');
         if ($operation == 'del') {
             // If operation is delete , it will delete, otherwise edit it
             ParticipantShare::model()->deleteRow($shareIds);
